Fix for inability to export via FBX
On latest 2.8-beta build (tested on Windows), I get an error if I try to export FBX.

Looks like ds_sp_obj_export_sel(...) was updated for 2.8 in commit tagged 1_5_0, but the same change to similar function ds_sp_fbx_export_sel(...) was overlooked.  I didn't look for other issues, but this fix was enough to get exporting to work for me.

Side note:  I'd recommend refactoring out the common parts between the OBJ and FBX code paths into common functions to avoid missing things like that in the future.  (or just have a "exportFileType" parameter that you pass in)